Finished Tile

+++ MCO1/src/game/crop/HarvestCropReport.java
@@ -4,7 +4,7 @@ public class HarvestCropReport {
     private boolean success = true;
     private String message = "Harvest successful!";
     private int productsProduced;
-    private int profit; // money gained from harvest
+    private double profit; // money gained from harvest
     private double expGained;
 
 
@@ -21,35 +21,43 @@ public class HarvestCropReport {
         this.profit = profit;
         this.expGained = expGained;
     }
-    
 
     public boolean isSuccess() {
         return success;
     }
+
     public void setSuccess(boolean success) {
         this.success = success;
     }
+
     public String getMessage() {
         return message;
     }
+
     public void setMessage(String message) {
         this.message = message;
     }
+
     public int getProductsProduced() {
         return productsProduced;
     }
+
     public void setProductsProduced(int productsProduced) {
         this.productsProduced = productsProduced;
     }
-    public int getProfit() {
+
+    public double getProfit() {
         return profit;
     }
-    public void setProfit(int profit) {
+
+    public void setProfit(double profit) {
         this.profit = profit;
     }
+
     public double getExpGained() {
         return expGained;
     }
+
     public void setExpGained(double expGained) {
         this.expGained = expGained;
     }
